requirements.txt
Django
python-decouple
python-magic
django-crispy-forms
crispy-bootstrap5
django-el-pagination
celery
django-celery-results
redis
humanize
astropy_healpix
mocpy
